write not done
Read's seem to be working fine, but the write ISR still needs work. It
appears that there is a problem with CS not being asserted and negated
properly. This may be due to the locations of asserts and negations but
also may be due to an imbalance in how the SPI1BUF is read and written.